+++ packages/dashboard-pages/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @getmunin/dashboard-pages
 
+## 4.31.0
+
+### Patch Changes
+
+- @getmunin/types@4.31.0
+- @getmunin/ui@4.31.0
+
 ## 4.30.0
 
 ### Minor Changes